clay-voice/requirements.txt
# Clay Voice — Dependencies
# Install: pip install -r requirements.txt

# Core
anthropic>=0.89.0

# Speech-to-Text (local)
faster-whisper>=1.0.0

# Wake Word Detection (choose one)
# Option A: Picovoice (native Windows, custom wake word, free tier)
pvporcupine>=3.0.0
# Option B: openWakeWord (fully open-source)
# openwakeword>=0.6.0

# Text-to-Speech (local)
piper-tts>=1.2.0

# Audio I/O
sounddevice>=0.4.6
numpy>=1.24.0

# Memory & Embeddings
sentence-transformers>=2.2.0
sqlite-utils>=3.35

# Configuration
pyyaml>=6.0

# Optional: for Google Calendar context
# google-api-python-client>=2.0.0
# google-auth-oauthlib>=1.0.0
